+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The major version bump is due to dropped support for PHP `7.2` and a several bre
 * Added `Collection::last()` method.
 * Added support for the `samesite` cookie option (defaults to `Lax`).
 * Added `Query::withCountOf()` method to the ORM query builder.
+* Added abstract `AccessControlAllowOrigin` middleware.
 
 #### Changes